Add Exchange mailing list membership search
Exchange has these dynamic distribution lists that make it difficult
to determine the full membership of the mailing list, particularly as
updating the memberOf attribute is apparently a non-automatic job, so
these are not often in-sync.

This tool recursively builds up the full membership list of an
Exchange mailing list / distribution group.

This code was written by Ed Fine, though he disclaims his usual
standards of quality as this is a quick one-off, though the code is
really still plenty nice.

Commit-message-by: Travis Cross <tc@traviscross.com>
Signed-off-by: Travis Cross <tc@traviscross.com>
Signed-off-by: Edwin Fine <emofine@gmail.com>